customer retrieval for order creation
