+++ src/main/java/bo/edu/ucb/syntax_flavor_backend/user/entity/User.java
@@ -9,16 +9,12 @@ package bo.edu.ucb.syntax_flavor_backend.user.entity;
 import java.io.Serializable;
 import java.util.Collection;
 import java.util.Date;
-
-import bo.edu.ucb.syntax_flavor_backend.order.entity.Order;
 import jakarta.persistence.Basic;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.NamedQueries;
 import jakarta.persistence.NamedQuery;
 import jakarta.persistence.OneToMany;
@@ -31,14 +27,16 @@ import jakarta.persistence.TemporalType;
  * @author Usuario
  */
 @Entity
-@Table(name = "customers")
+@Table(name = "users")
 @NamedQueries({
-    @NamedQuery(name = "Customers.findAll", query = "SELECT c FROM Customers c"),
-    @NamedQuery(name = "Customers.findById", query = "SELECT c FROM Customers c WHERE c.id = :id"),
-    @NamedQuery(name = "Customers.findByNit", query = "SELECT c FROM Customers c WHERE c.nit = :nit"),
-    @NamedQuery(name = "Customers.findByCreatedAt", query = "SELECT c FROM Customers c WHERE c.createdAt = :createdAt"),
-    @NamedQuery(name = "Customers.findByUpdatedAt", query = "SELECT c FROM Customers c WHERE c.updatedAt = :updatedAt")})
-public class Customers implements Serializable {
+    @NamedQuery(name = "Users.findAll", query = "SELECT u FROM Users u"),
+    @NamedQuery(name = "Users.findById", query = "SELECT u FROM Users u WHERE u.id = :id"),
+    @NamedQuery(name = "Users.findByName", query = "SELECT u FROM Users u WHERE u.name = :name"),
+    @NamedQuery(name = "Users.findByEmail", query = "SELECT u FROM Users u WHERE u.email = :email"),
+    @NamedQuery(name = "Users.findByPhone", query = "SELECT u FROM Users u WHERE u.phone = :phone"),
+    @NamedQuery(name = "Users.findByCreatedAt", query = "SELECT u FROM Users u WHERE u.createdAt = :createdAt"),
+    @NamedQuery(name = "Users.findByUpdatedAt", query = "SELECT u FROM Users u WHERE u.updatedAt = :updatedAt")})
+public class User implements Serializable {
 
     private static final long serialVersionUID = 1L;
     @Id
@@ -46,25 +44,34 @@ public class Customers implements Serializable {
     @Basic(optional = false)
     @Column(name = "id")
     private Integer id;
-    @Column(name = "nit")
-    private String nit;
+    @Basic(optional = false)
+    @Column(name = "name")
+    private String name;
+    @Basic(optional = false)
+    @Column(name = "email")
+    private String email;
+    @Column(name = "phone")
+    private String phone;
     @Column(name = "created_at")
     @Temporal(TemporalType.TIMESTAMP)
     private Date createdAt;
     @Column(name = "updated_at")
     @Temporal(TemporalType.TIMESTAMP)
     private Date updatedAt;
-    @OneToMany(mappedBy = "custom")
-    private Collection<Order> ordersCollection;
-    @JoinColumn(name = "users_id", referencedColumnName = "id")
-    @ManyToOne
-    private Users usersId;
+    @OneToMany(mappedBy = "usersId")
+    private Collection<Customer> customersCollection;
 
-    public Customers() {
+    public User() {
     }
 
-    public Customers(Integer id) {
+    public User(Integer id) {
         this.id = id;
+    }
+
+    public User(Integer id, String name, String email) {
+        this.id = id;
+        this.name = name;
+        this.email = email;
     }
 
     public Integer getId() {
@@ -75,12 +82,28 @@ public class Customers implements Serializable {
         this.id = id;
     }
 
-    public String getNit() {
-        return nit;
+    public String getName() {
+        return name;
     }
 
-    public void setNit(String nit) {
-        this.nit = nit;
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
     }
 
     public Date getCreatedAt() {
@@ -99,20 +122,12 @@ public class Customers implements Serializable {
         this.updatedAt = updatedAt;
     }
 
-    public Collection<Order> getOrdersCollection() {
-        return ordersCollection;
+    public Collection<Customer> getCustomersCollection() {
+        return customersCollection;
     }
 
-    public void setOrdersCollection(Collection<Order> ordersCollection) {
-        this.ordersCollection = ordersCollection;
-    }
-
-    public Users getUsersId() {
-        return usersId;
-    }
-
-    public void setUsersId(Users usersId) {
-        this.usersId = usersId;
+    public void setCustomersCollection(Collection<Customer> customersCollection) {
+        this.customersCollection = customersCollection;
     }
 
     @Override
@@ -125,10 +140,10 @@ public class Customers implements Serializable {
     @Override
     public boolean equals(Object object) {
         // TODO: Warning - this method won't work in the case the id fields are not set
-        if (!(object instanceof Customers)) {
+        if (!(object instanceof User)) {
             return false;
         }
-        Customers other = (Customers) object;
+        User other = (User) object;
         if ((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id))) {
             return false;
         }
@@ -137,8 +152,7 @@ public class Customers implements Serializable {
 
     @Override
     public String toString() {
-        return "com.condominio.tables_syntax_flavor.Customers[ id=" + id + " ]";
+        return "com.condominio.tables_syntax_flavor.Users[ id=" + id + " ]";
     }
     
 }
-
